Refactors my shit old cum drip code into a status effect

+++ modular_sand/code/game/objects/effects/decals/cleanable/lewd_decals.dm
@@ -7,15 +7,26 @@
 	icon = 'modular_sand/icons/obj/genitals/effects.dmi'
 	icon_state = "drip1"
 	random_icon_states = list("drip1", "drip2", "drip3", "drip4")
-	var/total_amount = 1
 
 /obj/effect/decal/cleanable/semendrip/replace_decal(obj/effect/decal/cleanable/semendrip/C)
 	. = ..()
-	C.total_amount++
-	C.transfer_blood_dna(src.blood_DNA)
-	if(C.total_amount >= 10)
-		var/obj/effect/decal/cleanable/semen/S = new(loc)
-		S.transfer_blood_dna(C.blood_DNA)
+	reagents.trans_to(src, C.reagents.total_volume)
+	transfer_blood_dna(C.blood_DNA)
+	var/obj/effect/decal/cleanable/semen/S = (locate(/obj/effect/decal/cleanable/semen) in C.loc)
+	if(S)
+		C.reagents.trans_to(S, C.reagents.total_volume)
+		C.transfer_blood_dna(S.blood_DNA)
+		C.update_icon()
+		return
+	if(C.reagents.total_volume >= 10)
+		S = new(C.loc)
+		C.reagents.trans_to(S, C.reagents.total_volume)
+		C.transfer_blood_dna(S.blood_DNA)
+		S.update_icon()
 		qdel(C)
+	update_icon()
 
-	return TRUE
+/obj/effect/decal/cleanable/semendrip/update_icon()
+	. = ..()
+	add_atom_colour(blood_DNA_to_color(), FIXED_COLOUR_PRIORITY)
+	blend_mode = blood_DNA_to_blend()
